+++ src/app/actions/auth.ts
@@ -11,11 +11,11 @@ const schemaRegister = z.object({
         .regex(/^[a-zA-Z0-9]+$/, "Username must contain only alphabetic characters"),
     email: z.string().email(),
     password: z.string()
-    .min(6, { message: "Password must be at least 6 characters long" })
-    .regex(/[a-z]/, { message: "Password must contain at least one lowercase letter" })
-    .regex(/[A-Z]/, { message: "Password must contain at least one uppercase letter" })
-    .regex(/[0-9]/, { message: "Password must contain at least one digit" })
-    .regex(/[^a-zA-Z0-9]/, { message: "Password must contain at least one special character" }),
+        .min(6, { message: "Password must be at least 6 characters long" })
+        .regex(/[a-z]/, { message: "Password must contain at least one lowercase letter" })
+        .regex(/[A-Z]/, { message: "Password must contain at least one uppercase letter" })
+        .regex(/[0-9]/, { message: "Password must contain at least one digit" })
+        .regex(/[^a-zA-Z0-9]/, { message: "Password must contain at least one special character" }),
 
     firstName: z.string().min(3).regex(/^[a-zA-Z]+$/, "Username must contain only alphabetic characters"),
     lastName: z.string().min(3).regex(/^[a-zA-Z]+$/, "Username must contain only alphabetic characters"),
@@ -24,10 +24,10 @@ const schemaRegister = z.object({
 }).refine(data => data.password === data.passwordReEnter, {
     message: "Passwords must match",
     path: ["passwordReEnter"],
-  });
+});
 
 
-export async function register(prevState: any, formData: FormData) { 
+export async function register(prevState: any, formData: FormData) {
 
     const validatedFields = schemaRegister.safeParse({
         username: formData.get('username'),
@@ -36,8 +36,8 @@ export async function register(prevState: any, formData: FormData) {
         firstName: formData.get('firstName'),
         lastName: formData.get('lastName'),
         passwordReEnter: formData.get('passwordReEnter')
-            
-    }) 
+
+    })
 
     if (!validatedFields.success) {
         return {
@@ -61,8 +61,8 @@ export async function register(prevState: any, formData: FormData) {
                 'Content-Type': 'application/json',
             },
         });
+        console.log(response)
         const result: any = await response.json();
-
         if (!result.success && result.errors) {
             const jsonObject = JSON.parse(result.errors);
 
@@ -71,10 +71,10 @@ export async function register(prevState: any, formData: FormData) {
             }
         }
         return result;
-    } 
+    }
 }
 export async function deleteAccount(prevState: any, formData: FormData) {
-  
+
     const session = await getServerSession(authOptions)
     if (!session || !session.user) {
         return {
@@ -89,6 +89,7 @@ export async function deleteAccount(prevState: any, formData: FormData) {
             'Authorization': `Bearer ${session.accessToken}`,
         },
     });
+
     const result: any = await response.json();
 
     if (!result.success && result.errors) {